Move the `ui-components` module under the `@bayou` namespace.
