+++ local-modules/@bayou/ui-components/Avatars/avatar.module.less
