+++ local-modules/@bayou/ui-components/SharingStatus/sharing-status.module.less
