+++ local-modules/@bayou/ui-components/Star/index.js
