+++ local-modules/@bayou/ui-components/Star/star.module.less
