+++ local-modules/doc-client/package.json
@@ -18,7 +18,7 @@
     "@bayou/state-machine": "local",
     "@bayou/typecheck": "local",
     "@bayou/ui-embeds": "local",
-    "ui-components": "local",
+    "@bayou/ui-components": "local",
     "@bayou/util-client": "local",
     "@bayou/util-common": "local"
   }
